+++ src/array/codec/array_to_bytes.rs
@@ -3,8 +3,9 @@
 pub mod bytes;
 pub mod codec_chain;
 
+#[cfg(feature = "pcodec")]
+pub mod pcodec;
 #[cfg(feature = "sharding")]
 pub mod sharding;
-
 #[cfg(feature = "zfp")]
 pub mod zfp;